+++ src/declarations.d.ts
@@ -1,0 +1,2 @@
+declare module 'is-json';
+declare module 'is-base64';